requirements_dev.txt
bgpy_pkg[test]==8.1.1